+++ jib-core/CHANGELOG.md
@@ -5,7 +5,14 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Multiple additions to `ImageReference` to separate `tag` and `digest`. ([#1481](https://github.com/GoogleContainerTools/jib/issues/1481))
+    - `of(registry, repository, tag, digest)` to create an image from a digest
+    - `isValidDigest(digest)` to check if a string is a valid digest
+    - `getDigest()` to get the digest
+    - `parse()` now supports image reference strings containing both a tag and a digest
+
 ### Changed
+- `ImageReference#isTagDigest` is now deprecated.
 
 ### Fixed